Refactor code base structure to segregate domain and application logic.
