+++ core/src/main/java/org/kritiniyoga/karmayoga/core/domain/entities/Schedule.java
@@ -1,0 +1,15 @@
+package org.kritiniyoga.karmayoga.core.domain.entities;
+
+import java.util.UUID;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import org.kritiniyoga.karmayoga.core.domain.values.TimeSlot;
+
+
+@Data
+@AllArgsConstructor
+public class Schedule {
+  UUID id;
+  Task task;
+  TimeSlot slot;
+}
