--- core/src/main/java/org/kritiniyoga/karmayoga/core/entities/Schedule.java
@@ -1,13 +1,0 @@
-package org.kritiniyoga.karmayoga.core.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import org.kritiniyoga.karmayoga.core.values.TimeSlot;
-
-
-@Data
-@AllArgsConstructor
-public class Schedule {
-    Task task;
-    TimeSlot slot;
-}
