--- core/src/main/java/org/kritiniyoga/karmayoga/core/values/Priority.java
@@ -1,5 +1,0 @@
-package org.kritiniyoga.karmayoga.core.values;
-
-public enum Priority {
-    LOW, MEDIUM, HIGH
-}
