+++ core/src/test/java/org/kritiniyoga/karmayoga/core/domain/factories/TimeSlotCreation.java
@@ -1,8 +1,8 @@
-package org.kritiniyoga.karmayoga.core.factories;
+package org.kritiniyoga.karmayoga.core.domain.factories;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.kritiniyoga.karmayoga.core.values.TimeSlot;
+import org.kritiniyoga.karmayoga.core.domain.values.TimeSlot;
 
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
